+++ src/modules/account/account.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { JwtModule } from '@nestjs/jwt';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { jwtConstants } from 'src/modules/auth/constants';
+import { jwtConstants } from 'src/modules/authentication/constants';
 import { AccountController } from './account.controller';
 import { AccountService } from './account.service';
 import {Pet, PetOwner, PetSitter} from 'src/entities'
