+++ src/modules/index.ts
@@ -1,5 +1,5 @@
 export * from './account/account.module'
-export * from './auth/auth.module'
+export * from './authentication/auth.module'
 export * from './booking/booking.module'
 export * from './dummy/dummy.module'
 export * from './notification/notification.module'
